fix inconsistency in Join logic; some proofs
Not a bug, but the sequence of lookup messages in Join leaves
cur_request unchanged. This means that timeouts_in would give the wrong
response if we tried to use end_query. We don't: the Join code
hard-codes its cleared and set timeouts.

Also: prove that requests get responses at recv_handler.